requirements_linux.txt
Pillow>=9.0.0
pyinstaller>=5.0.0
# Additional Linux dependencies that might be needed
# These are system packages that should be installed via package manager:
# sudo apt-get install python3-dev python3-tk tk-dev
# sudo apt-get install libjpeg-dev libpng-dev libtiff-dev libwebp-dev 